devices: window_wsl: add stub for windows WSL device
This is Linux running on WSL on a Windows machine. We will add specific
things for WSL in this class, such as running cmd.exe and other misc
items as needed (not in this commit).

For this to work one needs to start a ssh server inside WSL linux system
(ideally debian based). One needs to verify the default route goes
through the DUT by changing the metric on the interfaces and routes such
that this is the case. Also, make sure things like the firewall are
turned off.

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>